requirements.txt
Flask~=3.0.3
requests~=2.32.3
APScheduler~=3.10.4
python-dotenv~=1.0.1